Texture Revisions, Rarepons, Item Icons (Unfinished)
Added bird rarepon line, item icons, and revised some mirrored textures. The item icons are not finished yet, the ones that are not done have their original texture still in place. Usually I don't like doing this, but I've had the spritesheet lying around for half of a year now.